fixed authMiddleware issues.

+++ src/services/contactsService.js
@@ -1,11 +1,20 @@
+const jwt = require("jsonwebtoken");
 const { Contact } = require("../db/contactModel");
 const { WrongParamsError, NotAuthorizedError } = require("../helpers/errors");
 
-const getContacts = async (userId) => {
+const getContacts = async (token) => {
+  if (!token) throw new NotAuthorizedError("Authorization error");
+
+  const { _id: userId } = jwt.verify(token, process.env.JWT_SECRET);
   const contacts = await Contact.find({ userId });
+
   return contacts;
 };
-const getContactById = async (contactId, userId) => {
+
+const getContactById = async (contactId, token) => {
+  if (!token) throw new NotAuthorizedError("Authorization error");
+
+  const { _id: userId } = jwt.verify(token, process.env.JWT_SECRET);
   const contact = await Contact.findOne({ _id: contactId, userId });
 
   if (!contact) {
@@ -13,13 +22,22 @@ const getContactById = async (contactId, userId) => {
   }
   return contact;
 };
-const addContact = async ({ name, email, phone, favorite }, userId) => {
-  console.log(name);
+
+const addContact = async ({ name, email, phone, favorite }, token) => {
+  if (!token) throw new NotAuthorizedError("Authorization error");
+
+  const { _id: userId } = jwt.verify(token, process.env.JWT_SECRET);
   const contact = new Contact({ name, email, phone, favorite, userId });
+
   await contact.save();
 };
-const removeContact = async (contactId, userId) => {
+
+const removeContact = async (contactId, token) => {
+  if (!token) throw new NotAuthorizedError("Authorization error");
+
+  const { _id: userId } = jwt.verify(token, process.env.JWT_SECRET);
   const contact = await Contact.findOneAndDelete({ _id: contactId, userId });
+
   if (!contact) {
     throw new WrongParamsError(`Contact with id ${contactId} can't be found`);
   }
@@ -27,8 +45,12 @@ const removeContact = async (contactId, userId) => {
 const updateContact = async (
   contactId,
   { name, email, phone, favorite },
-  userId
+  token
 ) => {
+  if (!token) throw new NotAuthorizedError("Authorization error");
+
+  const { _id: userId } = jwt.verify(token, process.env.JWT_SECRET);
+
   await Contact.findOneAndUpdate(
     { _id: contactId, userId },
     {
